@@ -46,47 +46,48 @@
     "@rjsf/core": "^5.24.13",
     "@rjsf/utils": "^5.24.13",
     "@rjsf/validator-ajv8": "^5.24.13",
-    "@sentry/nextjs": "^10.17.0",
+    "@sentry/nextjs": "^10.19.0",
     "@types/json-schema": "^7.0.15",
     "@uppy/aws-s3": "^5.0.1",
-    "@uppy/core": "^5.0.2",
+    "@uppy/core": "^5.1.0",
     "@uppy/dashboard": "^5.0.2",
     "@uppy/react": "^5.1.0",
     "bcrypt-ts": "^7.1.0",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "cmdk": "^1.1.1",
-    "core-js": "^3.45.1",
+    "core-js": "^3.46.0",
     "dayjs": "^1.11.18",
     "js-base64": "^3.7.8",
     "js-sha256": "^0.11.1",
     "lucide-react": "^0.544.0",
     "next": "^13.5.11",
+    "next-seo": "^6.8.0",
     "nextjs-google-analytics": "^2.3.7",
-    "nextra": "^2.13.2",
-    "nextra-theme-docs": "^2.13.4",
+    "nextra": "^3.3.1",
+    "nextra-theme-docs": "^3.3.1",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "react-hook-form": "^7.63.0",
+    "react-hook-form": "^7.64.0",
     "sha512crypt-node": "^1.0.2",
     "sonner": "^2.0.7",
     "strip-indent": "^4.1.0",
     "swr": "^2.3.6",
     "tailwind-merge": "^1.14.0",
     "tailwindcss-animate": "^1.0.7",
-    "zod": "^4.1.11"
+    "zod": "^4.1.12"
   },
   "devDependencies": {
     "@next/bundle-analyzer": "^14.2.33",
     "@types/jest": "^30.0.0",
     "@types/node": "24.6.0",
-    "@types/react": "^18.3.25",
+    "@types/react": "^18.3.26",
     "ajv-cli": "^5.0.0",
     "ajv-formats": "^3.0.1",
     "autoprefixer": "^10.4.21",
     "axios": "^1.12.2",
     "dedent": "^1.7.0",
-    "eslint": "^9.36.0",
+    "eslint": "^9.37.0",
     "eslint-config-next": "^15.5.4",
     "eslint-config-prettier": "^10.1.8",
     "eslint-plugin-mdx": "^3.6.2",
@@ -98,7 +99,7 @@
     "quicktype": "^23.2.6",
     "sharp-cli": "^5.2.0",
     "slugify": "^1.6.6",
-    "tailwindcss": "^3.4.17",
+    "tailwindcss": "^3.4.18",
     "ts-jest": "^29.4.4",
     "typescript": "^5.9.3"
   },
