@@ -1,4 +1,4 @@
-{
+export default {
     "*": {
         "display": "hidden",
         "theme": {
@@ -8,4 +8,4 @@
             "typesetting": "article"
         }
     }
-}
+};
