@@ -1,4 +1,4 @@
-{
+export default {
     "overview": {
         "display": "hidden"
     },
