@@ -1,4 +1,4 @@
-{
+export default {
   "*": {
     "theme": {
       "timestamp": false
