@@ -1,4 +1,4 @@
-{
+export default {
     "index": "Introduction",
     "sponsor": "Sponsor",
     "join": "Join WATcloud"
